feat(cli): allow to modify cleartext setting from capacitor.config.json (#2397)
