+++ cli/src/android/update.ts
@@ -1,7 +1,7 @@
 import { Config } from '../config';
-import { buildXmlElement, checkPlatformVersions, logFatal, logInfo, parseXML, resolveNode, runTask } from '../common';
+import { buildXmlElement, checkPlatformVersions, logFatal, resolveNode, runTask } from '../common';
 import { getAndroidPlugins } from './common';
-import { checkAndInstallDependencies, handleCordovaPluginsJS } from '../cordova';
+import { checkAndInstallDependencies, handleCordovaPluginsJS, writeCordovaAndroidManifest } from '../cordova';
 import { convertToUnixPath, copySync, readFileAsync, removeSync, writeFileAsync} from '../util/fs';
 import { join, relative, resolve } from 'path';
 import { Plugin, PluginType, getAllElements, getFilePath, getPlatformElement, getPluginPlatform, getPluginType, getPlugins, printPlugins } from '../plugin';
@@ -32,7 +32,7 @@ export async function updateAndroid(config: Config) {
   await handleCordovaPluginsJS(cordovaPlugins, config, platform);
   await installGradlePlugins(config, capacitorPlugins, cordovaPlugins);
   await handleCordovaPluginsGradle(config, cordovaPlugins);
-  await writeCordovaAndroidManifest(cordovaPlugins, config);
+  await writeCordovaAndroidManifest(cordovaPlugins, config, platform);
 
   const incompatibleCordovaPlugins = plugins
   .filter(p => getPluginType(p, platform) === PluginType.Incompatible);
@@ -200,76 +200,4 @@ async function getPluginsTask(config: Config) {
     const androidPlugins = getAndroidPlugins(allPlugins);
     return androidPlugins;
   });
-}
-
-async function writeCordovaAndroidManifest(cordovaPlugins: Plugin[], config: Config) {
-  const pluginsFolder = resolve(config.app.rootDir, 'android', config.android.assets.pluginsFolderName);
-  const manifestPath = join(pluginsFolder, 'src', 'main', 'AndroidManifest.xml');
-  let rootXMLEntries: Array<any> = [];
-  let applicationXMLEntries: Array<any> = [];
-  let applicationXMLAttributes: Array<any> = [];
-  cordovaPlugins.map(async p => {
-    const editConfig = getPlatformElement(p, platform, 'edit-config');
-    const configFile = getPlatformElement(p, platform, 'config-file');
-    editConfig.concat(configFile).map(async (configElement: any) => {
-      if (configElement.$ && (configElement.$.target && configElement.$.target.includes('AndroidManifest.xml') || configElement.$.file && configElement.$.file.includes('AndroidManifest.xml'))) {
-        const keys = Object.keys(configElement).filter(k  => k !== '$');
-        keys.map(k => {
-          configElement[k].map((e: any) => {
-            const xmlElement = buildXmlElement(e, k);
-            const pathParts = getPathParts(configElement.$.parent || configElement.$.target);
-            if (pathParts.length > 1) {
-              if (pathParts.pop() === 'application') {
-                if (configElement.$.mode && configElement.$.mode === 'merge') {
-                  Object.keys(e.$).map((ek: any) => {
-                    applicationXMLAttributes.push(`${ek}="${e.$[ek]}"`);
-                  });
-                } else if (!applicationXMLEntries.includes(xmlElement) && !contains(applicationXMLEntries, xmlElement, k)) {
-                  applicationXMLEntries.push(xmlElement);
-                }
-              } else {
-                logInfo(`plugin ${p.id} requires to add \n  ${xmlElement} to your AndroidManifest.xml to work`);
-              }
-            } else {
-              if (!rootXMLEntries.includes(xmlElement) && !contains(rootXMLEntries, xmlElement, k)) {
-                rootXMLEntries.push(xmlElement);
-              }
-            }
-          });
-        });
-      }
-    });
-  });
-  let content = `<?xml version='1.0' encoding='utf-8'?>
-<manifest package="capacitor.android.plugins"
-xmlns:android="http://schemas.android.com/apk/res/android"
-xmlns:amazon="http://schemas.amazon.com/apk/res/android">
-<application ${applicationXMLAttributes.join('\n')}>
-${applicationXMLEntries.join('\n')}
-</application>
-${rootXMLEntries.join('\n')}
-</manifest>`;
-  content = content.replace(new RegExp(('$PACKAGE_NAME').replace('$', '\\$&'), 'g'), config.app.appId);
-  await writeFileAsync(manifestPath, content);
-}
-
-function getPathParts(path: string) {
-  const rootPath = 'manifest';
-  path = path.replace('/*', rootPath);
-  let parts = path.split('/').filter(part => part !== '');
-  if (parts.length > 1 || parts.includes(rootPath)) {
-    return parts;
-  }
-  return [rootPath, path];
-}
-
-function contains(a: Array<any>, obj: any, k: string) {
-  const element = parseXML(obj);
-  for (var i = 0; i < a.length; i++) {
-    const current = parseXML(a[i]);
-    if (element && current && current[k]  && element[k] && current[k].$ && element[k].$ && element[k].$['android:name'] === current[k].$['android:name']) {
-      return true;
-    }
-  }
-  return false;
 }
