+++ cli/src/config.ts
@@ -90,6 +90,9 @@ export class Config implements CliConfig {
       templateName: 'app-template',
       templateDir: '',
       pluginsTemplateDir: ''
+    },
+    server: {
+      cleartext: false
     }
   };
